+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Changes to this project are documented in this file. More detail and links can b
 - AUTOTOOL-1019: Splunk legacy tmstats - add tenant and application data
 - AUTOTOOL-1128: Declarations with large secrets may timeout
 - AUTOTOOL-1154: Passphrases should be obfuscated in consumer trace files
+- AUTOTOOL-1147: Add 'profiles' data (profiles attached to Virtual Server) to 'virtualServers'
 - AUTOTOOL-896: [GitHub #26](https://github.com/F5Networks/f5-telemetry-streaming/pull/26): Use baseMac instead of hostname to fetch CM device
 - AUTOTOOL-1160: cipherText validation when protected by SecureVault
 ### Changed